add: tutorial mixin
