+++ pages/index.vue
@@ -7,6 +7,7 @@
     <div id="operation-area-wrapper">
       <operation-area />
     </div>
+
     <player />
   </div>
 </template>
@@ -19,9 +20,10 @@ import OperationArea from '@/components/operationArea.vue'
 import Player from '@/components/Player.vue'
 import { firebaseAuth, firestoreAccessor } from '@/plugins/firebase'
 import { Music } from '@/types/music'
+import lesson from '@/mixins/lesson.ts'
 
 type DataType = {
-  dialog: boolean
+  lessonAvailable: boolean
 }
 
 export default {
@@ -30,6 +32,7 @@ export default {
     OperationArea,
     Player
   },
+  mixins: [lesson],
   async fetch({ route, store }: Context) {
     firebaseAuth.onAuthStateChanged((user) => {
       if (user) {
@@ -63,59 +66,7 @@ export default {
   },
   data(): DataType {
     return {
-      dialog: false
-    }
-  },
-  mounted() {
-    this.$driver.options.allowClose = false
-    this.$driver.options.opacity = 0
-    this.$driver.options.nextBtnText = 'OK!'
-    this.$driver.options.prevBtnText = '戻る'
-    this.$driver.options.closeBtnText = 'やめる'
-    this.$driver.options.doneBtnText = '終わる'
-    this.$driver.options.showButtons = true
-
-    this.$driver.options.onNext = (step) => {
-      if (step && step.options.delay) {
-        const currentIndex = this.$driver.currentStep
-        this.$driver.preventMove()
-        this.$driver.reset()
-        setTimeout(() => {
-          this.setSteps()
-          this.$driver.start(currentIndex + 1)
-        }, step.options.delay)
-      }
-    }
-
-    this.$driver.options.onHighlightStarted = (step) => {
-      const nextEvent = step.options.nextEvent
-      if (nextEvent) {
-        const elem = document.getElementById(nextEvent.elemId.slice(1))
-        if (!elem) {
-          return
-        }
-        const f = () => {
-          this.$driver.handleNext()
-          elem.removeEventListener('click', f, false)
-        }
-        elem.addEventListener('click', f, false)
-      }
-    }
-
-    this.$driver.options.onHighlighted = (step) => {
-      if (step.options.disableTouch) {
-        step.node.classList.remove('driver-highlighted-element')
-      }
-    }
-
-    this.setSteps()
-    this.$driver.start()
-  },
-  methods: {
-    setSteps() {
-      this.$driver.defineSteps(
-        this.$accessor.tutorial.steps.map((step, index) => ({ ...step, index }))
-      )
+      lessonAvailable: true
     }
   }
 }
